add validateAsynch and checkAsynch
When excessively validating properties and requireing IO operations for some validations, asynchronous calls can be very handy.
The implemented calls validateAsynch and checkAsynch have the same signatures as their synchronous counterparts, but return
a java.util.concurrent.ComplitableFuture instead of void or boolean.

validateAsynch return a future which returns the actual value again on calling get().
checkAsynch returns a future which return a boolean value (if valid or not) on calling get().

When the error handling action throws an exception, it can either be handlet when calling get() and catching the ExecutionException,
its cause attribute will be the e.g. ValidationException, or use the handle/handleAsynch continouation functions.

Further tasks in this branch:
- clarify state issues on Validator instances and ValidcoolConfiguration
- should we wrap completable future?
- combining validations